+++ test/transform/ElementUtilities.js
@@ -1,6 +1,6 @@
 import assert from 'assert'
-import fixtureIO from './utilities/FixtureIO'
-import pagelib from '../build/wikimedia-page-library-transform'
+import fixtureIO from '../utilities/FixtureIO'
+import pagelib from '../../build/wikimedia-page-library-transform'
 
 const elementUtilities = pagelib.test.ElementUtilities
 let document